% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSetExperiment.R
\name{FileSetExperiment}
\alias{FileSetExperiment}
\title{A FileSetExperiment constructor.}
\usage{
FileSetExperiment(fset, ...)
}
\arguments{
\item{fset}{a fileSet object (really just a list)}
}
\description{
A FileSetExperiment constructor.
}
