% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDATstubs.R
\name{getIDATstubs}
\alias{getIDATstubs}
\title{Given a list of URLs for IDAT files, extract basenames (i.e., ABC123_R0XC0Y)}
\usage{
getIDATstubs(x)
}
\arguments{
\item{x}{vector of URLs/paths or a SummarizedExperiment-like object}
}
\value{
vector of IDAT stubs or a SummarizedExperiment-like object
}
\description{
For a GEO dataset or an SE made from GEOquery results, this will usually
be found in the $supplementary_file column of the sample data or colData.
If a SummarizedExperiment-like object is provided, it will be returned with
a new colData column 'Basename' holding the resulting IDAT file stubs.
}
